fix(14148): refactored style location
