--- src/components/LHNOptionsList/styles.native.js
@@ -1,10 +1,0 @@
-import styles from '../../styles/styles';
-
-const compactAlternateTextStyle = {};
-
-const compactContentContainerStyles = styles.alignItemsCenter;
-
-export {
-    compactAlternateTextStyle,
-    compactContentContainerStyles,
-};
